+++ frontend/src/App.js
@@ -9,28 +9,6 @@ import PrivateRoute from "./PrivateRoute";
 function App() {
     const [jwt, setJwt] = useLocalState("","jwt");
 
-    useEffect(() => {
-        if(!jwt){
-            const reqBody = {
-                username: "petowner",
-                password: "asdfasdf",
-            };
-            fetch("api/auth/login", {
-                headers: {
-                    "Content-Type": "application/json"
-                },
-                method: "post",
-                body: JSON.stringify(reqBody),
-            })
-                .then((response) => Promise.all([response.json(), response.headers]))
-                .then(([body, headers]) => {
-                        setJwt(headers.get("authorization"));
-                    });
-        }
-    }, []);
-
-
-
     return(
         <Routes>
             <Route path="/dashboard" element={
